Refactor: move EmailSender to utils
